Fixes and standarized builder arguments.
