+++ packages/chan/src/commands/init.js
@@ -1,13 +1,11 @@
 import { resolve } from 'path'
-import toVFile from 'to-vfile'
 import boxen from 'boxen'
-
 import { promises as fs } from 'fs'
 
 import { initialize } from '@geut/chan-core'
 
 import { createLogger } from '../logger.js'
-import { write } from '../write.js'
+import { read, write } from '../vfs.js'
 
 export const command = 'init [dir]'
 export const description = 'Initialize CHANGELOG.md file'
@@ -17,8 +15,8 @@ export const builder = {
     alias: ['p', 'path'],
     default: '.'
   },
-  o: {
-    alias: 'overwrite',
+  overwrite: {
+    alias: 'o',
     describe: 'Overwrite the current CHANGELOG.md',
     type: 'boolean',
     default: false
@@ -29,7 +27,7 @@ export async function handler ({ dir, overwrite, verbose, stdout }) {
   const { report, success, info } = createLogger({ scope: 'init', verbose, stdout })
 
   try {
-    const file = await readFile(resolve(dir, 'CHANGELOG.md'))
+    const file = await read(resolve(dir, 'CHANGELOG.md'))
 
     await initialize(file, { overwrite: overwrite || stdout })
 
@@ -47,18 +45,5 @@ export async function handler ({ dir, overwrite, verbose, stdout }) {
     info('Update the npm script `version` in your package.json to release automatically:')
     console.log(boxen('chan release ${npm_package_version} && git add .', { padding: 1, float: 'center' })) // eslint-disable-line no-template-curly-in-string
   } catch (err) {
-  }
-}
-
-async function readFile (path) {
-  try {
-    const file = await toVFile.read(path)
-    return file
-  } catch (err) {
-    if (err.code === 'ENOENT') {
-      return toVFile(path)
-    } else {
-      throw err
-    }
   }
 }
