--- packages/chan/src/write.js
@@ -1,9 +1,0 @@
-import toVFile from 'to-vfile'
-
-export async function write ({ file, stdout }) {
-  if (stdout) {
-    process.stdout.write(file.toString())
-  } else {
-    await toVFile.write(file)
-  }
-}
